Add renames for Plone 6.0 and 6.1.
They are for the following interfaces:

- `plone.restapi.behaviors.ITiles`
- `collective.dexteritytextindexer.behavior.IDexterityTextIndexer`
- `plone.app.discussion.behaviors.IAllowDiscussion`
- `plone.app.dexterity.behaviors.discussion.IAllowDiscussion`

On one Plone site on 6.0 I got these errors with `bin/instance zodbverify`.
The numbers here are the number of times the error was printed:

```
2507 AttributeError: module 'plone.restapi.behaviors' has no attribute 'ITiles'
12246 ModuleNotFoundError: No module named 'collective.dexteritytextindexer'
```

When I first migrate that site to 6.1, it becomes partially different:

```
2507 AttributeError: module 'plone.restapi.behaviors' has no attribute 'ITiles'
5055 ModuleNotFoundError: No module named 'plone.app.discussion.behaviors'
8694 ModuleNotFoundError: No module named 'collective.dexteritytextindexer'
```

For the record, the 6.1 site *does* have `plone.app.discussion` in its eggs, so that is not the reason for the extra error.
I actually could not find a version of `plone.app.discussion` which has a `behaviors` module, so I am not completely sure how this happened.
This site started out at Plone 5.1.  Anyway, the extra renames get rid of the error.

I have two other Plone 6.0 sites which do not need any of this.
Maybe they do need it when I migrate them to 6.1.